fix: schema uyumsuzlugu, DLQ eksikligi, yetki hatasi ve sayfalama sinirlari
- PasswordResetToken.id: Long -> Integer (DB migration SERIAL/integer ile
  uyumsuzdu, backend hic acilmiyordu)
- application.yml/SecurityConfig/OAuth2AuthenticationSuccessHandler: eski
  Vite portu (5173) varsayilanlari 3000'e cevrildi
- ExternalRabbitConfig: external.ingestion kuyrugunda DLQ ayari eksikti,
  hatali mesaj sonsuz dongude ayni kuyruga geri donebiliyordu
- MessageService.markAsRead: yetkisiz kullanicida sessizce 200 donuyordu,
  gercek 403'e cevrildi
- ChatController/FavoriteAdController: sayfalama size parametresine ust
  sinir eklendi (diger uclarla tutarli)

+++ src/main/java/com/works/patimati/repository/PasswordResetTokenRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PasswordResetTokenRepository extends JpaRepository<PasswordResetToken, Long> {
+public interface PasswordResetTokenRepository extends JpaRepository<PasswordResetToken, Integer> {
 
     // Frontend'den gelen token metnini veritabanında aramak için
     Optional<PasswordResetToken> findByToken(String token);
